+++ filestore/src/main/kotlin/dev/openrune/cache/filestore/definition/data/VarpType.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class VarpType(
     override var id: Int = -1,
-    var configType: Int = 0,
+    var configType: UShort = 0u,
     //Custom
     override var inherit: Int = -1
 ) : Definition
